Build 1.11
A few changes, but they are really effective. For one, I made a new board creation algorithm that is much faster. Basically, instead of starting over when messing something up, it goes to a previous iteration of the board when it was working. It may seem simple to me now, but earlier I seriously hadn't thought of it yet. I also replaced the zeroes with empty spaces in the rendering to cut down on clutter.

Things to work on:

- AI. Yes, I will work on it.
- Add more things to make the game seem more finished, perhaps a timer or a hint system, or difficulty. 
- I forgot what else I was going to do. Oh well!